removing some redundency, integrating plotlyviz nice features into default viz
